requirements.txt
tensorflow>=2.10
streamlit
pillow
numpy
matplotlib            
scikit-learn 